learning jdbc and flyway database migration

+++ src/main/java/com/harimahat/customer/CustomerRepository.java
@@ -6,8 +6,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
  * @Author hari.mahat on 28.7.2023
  * Project learn-spring3
  */
-public interface CustomerRepository extends JpaRepository<Customer, Integer> {
+public interface CustomerRepository extends JpaRepository<Customer, Long> {
 
         boolean existsCustomerByEmail(String email);
-        boolean existsCustomerById(Integer id);
+        boolean existsCustomerById(Long id);
 }
